pyproject.toml
[project]
name = "mcp-experiment-01"
version = "0.1.0"
description = "MCP server with web browsing and secure document reading tools"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "pyyaml>=6.0",
]

[project.scripts]
mcp-experiment-01 = "mcp_server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]
